requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
